feat: Add solution function for calculating binary string representation of operations
Add a new function called `solution` to the `operations.py` file in the `employee_register` directory. This function takes a list of operations and returns a binary string representation where 1 represents a possible operation and 0 represents an impossible operation. The function handles different types of operations, including adding obstacles and calculating the binary string based on the given operations.

Example usage of the `solution` function is also included at the end of the file.